Implement "params-inside-options" pattern for remaining surface area
Pave the way for merging in a branch from another fork that has not been
accepted upstream yet, which adds support for multiple azure storage
accounts, parameterised by call.

Previous commits in our fork here made breaking changes to a few
functions, e.g. `put_blob()`. Upstream, this function's final argument
is a keyword list defaulting to `[]`. In our fork, we've nested the
params keyword list inside another one, so that we can pass various
non-param overrides without having 2 keyword lists. The breaking change
isn't obvious in the type spec, but the user must pass a
differently-shaped keyword list.

The noaccOS fork has a similar motivation to add non-params overrides,
but they've taken the other approach: passing 2 keyword lists. When we
merge in their branch here, we'll have to pick a signature kind (2
keyword lists or a single nested one). Since we're already using a
nested one, we'll stick with that pattern and resolve the merge
conflicts.

First though, we must finish the interface change job that we'd started,
and nest params inside options for all functions. Interestingly, for
some functions, this is just a rename to make it clear we'd already
implicitly done this.

Co-authored-by: Harisudhan V R <harisudhan.v.r@maersk.com>